- Rename StructureArtifact to BaseArtifact - Move artifacts to core
