--- griptape/artifacts/__init__.py
@@ -1,12 +1,0 @@
-from griptape.artifacts.structure_artifact import StructureArtifact
-from griptape.artifacts.error_output import ErrorOutput
-from griptape.artifacts.text_output import TextOutput
-from griptape.artifacts.text_input import TextInput
-
-
-__all__ = [
-    "StructureArtifact",
-    "ErrorOutput",
-    "TextOutput",
-    "TextInput"
-]
