+++ griptape/core/artifacts/__init__.py
@@ -1,0 +1,12 @@
+from griptape.core.artifacts.base_artifact import BaseArtifact
+from griptape.core.artifacts.error_output import ErrorOutput
+from griptape.core.artifacts.text_output import TextOutput
+from griptape.core.artifacts.text_input import TextInput
+
+
+__all__ = [
+    "BaseArtifact",
+    "ErrorOutput",
+    "TextOutput",
+    "TextInput"
+]
