+++ griptape/core/artifacts/error_output.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 from typing import Optional
 from attr import define, field
-from griptape.artifacts import StructureArtifact
+from griptape.core.artifacts import BaseArtifact
 
 
 if TYPE_CHECKING:
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
 
 
 @define(frozen=True)
-class ErrorOutput(StructureArtifact):
+class ErrorOutput(BaseArtifact):
     value: Optional[str] = field()
     exception: Optional[Exception] = field(default=None, kw_only=True)
     task: Optional[BaseTask] = field(default=None, kw_only=True)
