+++ griptape/core/artifacts/text_input.py
@@ -1,11 +1,11 @@
 from typing import Optional
 from attr import define, field
-from griptape.artifacts import StructureArtifact
+from griptape.core.artifacts import BaseArtifact
 from griptape.core.tokenizers import BaseTokenizer
 
 
 @define(frozen=True)
-class TextInput(StructureArtifact):
+class TextInput(BaseArtifact):
     value: Optional[str] = field()
 
     def token_count(self, tokenizer: BaseTokenizer) -> Optional[int]:
