+++ griptape/core/drivers/prompt/hugging_face_hub_prompt_driver.py
@@ -1,7 +1,7 @@
 from attr import define, field, Factory
 from huggingface_hub import InferenceApi
 from transformers import AutoTokenizer
-from griptape.artifacts import TextOutput
+from griptape.core.artifacts import TextOutput
 from griptape.core.drivers import BasePromptDriver
 from griptape.core.tokenizers import HuggingFaceTokenizer
