+++ griptape/core/drivers/prompt/hugging_face_pipeline_prompt_driver.py
@@ -1,6 +1,6 @@
 from attr import define, field, Factory
 from transformers import pipeline, AutoTokenizer
-from griptape.artifacts import TextOutput
+from griptape.core.artifacts import TextOutput
 from griptape.core.drivers import BasePromptDriver
 from griptape.core.tokenizers import HuggingFaceTokenizer
