+++ tests/mocks/mock_value_driver.py
@@ -1,7 +1,7 @@
 from attr import define
 from griptape.core.drivers import BasePromptDriver
 from griptape.core.tokenizers import TiktokenTokenizer, BaseTokenizer
-from griptape.artifacts import TextOutput
+from griptape.core.artifacts import TextOutput
 
 
 @define
