+++ tests/unit/tasks/test_toolkit_task.py
@@ -1,5 +1,5 @@
 from tests.mocks.mock_tool.tool import MockTool
-from griptape.artifacts import ErrorOutput
+from griptape.core.artifacts import ErrorOutput
 from griptape.tasks import ToolkitTask, ToolSubtask
 from griptape.utils import ToolLoader
 from tests.mocks.mock_value_driver import MockValueDriver
